# api/internal/api/spec/openapi.yaml
openapi: 3.1.0
info:
  title: 75hard API
  version: "1.0.0"
  description: |
    Tracking API for the 75 Hard challenge: daily tasks, progress photos,
    meals and calories, workouts, meditation, Strava imports and body metrics.

    ## Authentication

    Every endpoint under `/api` except the auth and discovery routes needs a
    bearer token. Two kinds are accepted:

    - a **session JWT** from `POST /api/auth/login`
    - a **personal API token**, created in Settings, which looks like
      `75h_` followed by 64 hex characters

    Personal tokens carry scopes. A `read` token may only issue GET, HEAD and
    OPTIONS requests; anything that changes data needs `write`. A read-only
    token attempting a write gets 403, not 401 — the credential is valid, it
    simply is not permitted.

    ## Units

    Weight is always kilograms on the wire, regardless of the display
    preference on the account. Distances are metres, durations are seconds
    unless a field says minutes, and energy is kilocalories.

    ## Dates

    A day is the user's local calendar date, formatted `YYYY-MM-DD`. Day 1 of a
    program is its start date.
servers:
  - url: https://75hard.biswas.me
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /api/auth/signup:
    post:
      tags: [Auth]
      summary: Create an account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 12 }
                name: { type: string }
      responses:
        "201": { description: Account created, returns a token and the user }
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Sign in
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: |
            Either a session, or a two-factor challenge when the account needs
            a code. Check for `two_factor` before assuming sign-in finished:
            the challenge is signed with a different key and is not accepted
            anywhere else in this API.
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      token: { type: string }
                      user: { $ref: "#/components/schemas/User" }
                  - type: object
                    properties:
                      two_factor: { type: boolean, enum: [true] }
                      challenge:
                        type: string
                        description: Present it to /api/auth/2fa/verify with a code. Valid for five minutes.
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/auth/2fa/verify:
    post:
      tags: [Auth]
      summary: Finish a sign-in that stopped for a code
      description: |
        Accepts either a current authenticator code or an unused recovery code.
        A recovery code is spent the moment it works.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challenge, code]
              properties:
                challenge: { type: string }
                code: { type: string }
      responses:
        "200":
          description: A session token and the user
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/2fa/setup:
    post:
      tags: [Auth]
      summary: Generate a secret to enrol an authenticator app
      description: |
        Stores the secret unconfirmed and turns nothing on. Only a correct code
        from the app makes it a requirement, so a mis-scanned QR cannot lock the
        account out of itself.
      responses:
        "200":
          description: The secret and the otpauth URI to render as a QR code
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret: { type: string }
                  uri: { type: string }
        "409": { description: Two-factor is already on }
  /api/2fa/confirm:
    post:
      tags: [Auth]
      summary: Prove the secret was scanned and turn two-factor on
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string }
      responses:
        "200":
          description: |
            Enabled, with the recovery codes. They are stored hashed and are
            returned exactly once — there is no way to read them again.
          content:
            application/json:
              schema:
                type: object
                properties:
                  enabled: { type: boolean }
                  recovery_codes:
                    type: array
                    items: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/2fa/disable:
    post:
      tags: [Auth]
      summary: Turn two-factor off
      description: |
        Re-checks the person at the keyboard: a password for a password
        account, or a current code for one that signs in with Google.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password: { type: string }
                code: { type: string }
      responses:
        "200": { description: Turned off }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/passkeys/{passkeyID}:
    delete:
      tags: [Auth]
      summary: Remove a passkey
      parameters:
        - name: passkeyID
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Removed }
        "404": { description: No such passkey }
...
  /api/passkeys/register/finish:
    post:
      tags: [Auth]
      summary: Store a newly created passkey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id, credential]
              properties:
                session_id: { type: string }
                name: { type: string }
                credential: { type: object }
      responses:
        "200": { description: Registered }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/auth/passkeys/login/finish:
    post:
      tags: [Auth]
      summary: Complete a passwordless sign-in
      description: |
        No code is asked for afterwards. The authenticator has already checked
        a fingerprint, a face or a PIN before it would sign anything, so the
        assertion is itself two factors.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id, credential]
              properties:
                session_id: { type: string }
                credential: { type: object }
      responses:
        "200":
          description: A session token and the user
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/me:
    get:
      tags: [Auth]
      summary: The signed-in account
      responses:
        "200":
          description: The account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
    patch:
      tags: [Auth]
      summary: Update profile
      description: |
        `weight_unit` is a display preference only — weights are still sent and
        received in kilograms.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                timezone: { type: string, description: IANA name, e.g. America/Toronto }
                weight_unit: { type: string, enum: [kg, lb] }
      responses:
        "200":
          description: The updated account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
  /api/programs:
    get:
      tags: [Programs]
      summary: List attempts
      responses:
        "200":
          description: Every program, newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Program" }
    post:
      tags: [Programs]
      summary: Start a program
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                start_date: { type: string, format: date }
                length_days: { type: integer, default: 75 }
                daily_kcal_target: { type: integer, nullable: true }
      responses:
        "201":
          description: The new program
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Program" }
        "409": { description: An active program already exists }
  /api/programs/active:
    get:
      tags: [Programs]
      summary: The current attempt
      responses:
        "200":
          description: The active program
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Program" }
        "404": { description: No active program }
...
  /api/programs/{programID}/days:
    get:
      tags: [Days]
      summary: Every day, for the calendar
      parameters: [{ $ref: "#/components/parameters/ProgramID" }]
      responses:
        "200":
          description: One summary per day
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/DaySummary" }
  /api/programs/{programID}/days/{dayNumber}:
    get:
      tags: [Days]
      summary: One day in full
      parameters:
        - { $ref: "#/components/parameters/ProgramID" }
        - { $ref: "#/components/parameters/DayNumber" }
      responses:
        "200":
          description: The day
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Day" }
    patch:
      tags: [Days]
      summary: Save the note, weight or resting pulse
      description: |
        Weight is kilograms. Sending 0 for either number clears it rather than
        recording a zero reading.
      parameters:
        - { $ref: "#/components/parameters/ProgramID" }
        - { $ref: "#/components/parameters/DayNumber" }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                note: { type: string }
                weight_kg: { type: number, format: float }
                resting_hr: { type: integer, minimum: 25, maximum: 200 }
      responses:
        "200":
          description: The updated day
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Day" }
  /api/programs/{programID}/days/{dayNumber}/tasks/{taskID}:
    post:
      tags: [Days]
      summary: Tick a task, or record its value
      description: |
        `done` drives check, photo and text tasks. `value` drives number and
        duration tasks, which complete once the value reaches the target.
      parameters:
        - { $ref: "#/components/parameters/ProgramID" }
        - { $ref: "#/components/parameters/DayNumber" }
        - name: taskID
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                done: { type: boolean }
                value: { type: number }
                note: { type: string }
      responses:
        "200":
          description: The updated day
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Day" }
  /api/programs/{programID}/grid:
    get:
      tags: [Days]
      summary: A cell per activity per day
      parameters: [{ $ref: "#/components/parameters/ProgramID" }]
      responses:
        "200": { description: One row per task, one cell per day }
  /api/meals:
    post:
      tags: [Nutrition]
      summary: Log a meal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                day_number: { type: integer }
                name: { type: string }
                slot: { type: string, enum: [breakfast, lunch, dinner, snack] }
                kcal: { type: number }
                protein_g: { type: number }
                carbs_g: { type: number }
                fat_g: { type: number }
                photo_id: { type: integer, format: int64, nullable: true }
      responses:
        "201":
          description: The meal
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Meal" }
  /api/meals/{mealID}:
    patch:
      tags: [Nutrition]
      summary: Edit a meal
      parameters: [{ $ref: "#/components/parameters/MealID" }]
      responses:
        "200":
          description: The updated meal
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Meal" }
    delete:
      tags: [Nutrition]
      summary: Remove a meal
      parameters: [{ $ref: "#/components/parameters/MealID" }]
      responses:
        "200": { description: Removed }
  /api/meals/{mealID}/estimate:
    post:
      tags: [Nutrition]
      summary: Re-run a failed photo estimate
      description: |
        Food photos uploaded with `autolog=1` are estimated in the background.
        This retries one whose `estimate_status` is `failed` — usually because
        the daily AI limit was reached, which clears on its own.
      parameters: [{ $ref: "#/components/parameters/MealID" }]
      responses:
        "202": { description: Queued }
  /api/workouts:
    post:
      tags: [Workouts]
      summary: Log a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                day_number: { type: integer }
                kind: { type: string, enum: [indoor, outdoor] }
                activity: { type: string }
                minutes: { type: integer }
                kcal: { type: number, nullable: true }
                task_id:
                  type: integer
                  format: int64
                  description: Also credits this task with the total minutes logged
      responses:
        "201":
          description: The workout
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workout" }
  /api/workouts/{workoutID}:
    delete:
      tags: [Workouts]
      summary: Remove a session
      parameters:
        - name: workoutID
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200": { description: Removed }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        A session JWT, or a personal API token of the form
        `75h_` + 64 hex characters.
...
  responses:
    Unauthorized:
      description: Missing, invalid or expired credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...